agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "langchain>=1.2.13",
  "langchain-openai>=1.1.12",
  "langgraph>=0.4.19",
  "langgraph-cli[inmem]>=0.4.19",
  "pandas>=3.0.1",
  "pydantic>=2.12.5",
  "pyright>=1.1.408",
  "python-docx>=1.2.0",
  "python-dotenv>=1.2.2",
  "rapidfuzz>=3.0",
  "psycopg2-binary>=2.9",
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-multipart>=0.0.9",
]


[tool.pyright]
venvPath = "."
venv = ".venv"
